requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.3
django-admin-interface==0.29.2
django-allauth==65.2.0
django-colorfield==0.11.0
django-decouple==2.1
django-import-export==4.3.1
pillow==11.0.0
psycopg2-binary==2.9.10
python-slugify==8.0.4
sqlparse==0.5.2
tablib==3.7.0
text-unidecode==1.3
typing_extensions==4.12.2


# Time: 02:13:00